requirements.txt
PyQt6
reportlab
python-docx
docx2pdf
bcrypt
